Canvas mouse clicking is working X parameter type migration from int to float Add server checking if point is correct Refresh is working now

+++ src/main/java/app/model/Results.java
@@ -38,7 +38,7 @@ public class Results {
         for (int i = points.size() - 1; i >= 0; i--) {
             current = points.get(i);
             if (current.isCorrect()) {
-                int x = current.getX();
+                float x = current.getX();
                 float y = current.getY();
                 int r = current.getR();
                 String creationDateString = current.getCreationDateString();
